Correct 64bit pointer to 32bit integer casting problem ([https://trac.osgeo.org/fdo/ticket/273]).
git-svn-id: http://svn.osgeo.org/fdo/trunk@3800 d98eaee6-4212-0410-900c-dcc49bf69b99

+++ Fdo/Unmanaged/Src/Fdo/ClientServices/prefix.cpp
@@ -121,7 +121,7 @@ br_locate (void *symbol)
 
 	while (!feof (f))
 	{
-		unsigned int start, end;
+		unsigned long start, end;
 
 		if (!fgets (line, sizeof (line), f))
 			continue;
@@ -129,7 +129,7 @@ br_locate (void *symbol)
 			continue;
 
 		sscanf (line, "%x-%x ", &start, &end);
-		if (((unsigned int) symbol) >= start && ((unsigned int) symbol) < end)
+		if (((unsigned long) symbol) >= start && ((unsigned long) symbol) < end)
 		{
 			char *tmp;
 			size_t len;
